requirements.txt
tensorflow>=2.10
shap>=0.42
numpy>=1.23
pandas>=1.5
matplotlib>=3.6
Pillow>=9.0
scikit-learn>=1.1
jupyter>=1.0
